Adds bean validation to Question.
For that to work it was necessary to create a proper question controller with its own ModelAttributes. So all operations on questions moved out from FeedbackController to the newly created QuestionController.
To the DTO (or Command Object) validations were added, including a custom one.
This custom validation is obviously also implemented on this commit.